Start new region tracker classes.
The new region tracking classes are intended for tracking between images with
small baseline movements, such as video. To start this includes a classic KLT
tracker, and a pyramid tracker which composes with any other tracker. For
example, to make a pyramid KLT tracker, initialize the PyramidRegionTracker
with a KltRegionTracker.